# Fix coef and sign