--- a/lib/decimal.ex
+++ b/lib/decimal.ex
@@ -229,6 +229,11 @@ defmodule Decimal do
 
   def add(%Decimal{}, %Decimal{coef: :inf} = num2), do: num2
 
+  def add(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
+      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
+    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  end
+
   def add(%Decimal{} = num1, %Decimal{} = num2) do
     %Decimal{sign: sign1, coef: coef1, exp: exp1} = num1
     %Decimal{sign: sign2, coef: coef2, exp: exp2} = num2
@@ -308,6 +313,11 @@ defmodule Decimal do
 
   def compare(_num1, %Decimal{coef: :NaN} = num2),
     do: error(:invalid_operation, "operation on NaN", num2)
+
+  def compare(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
+      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
+    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  end
 
   def compare(%Decimal{} = num1, %Decimal{} = num2) do
     case sub(num1, num2) do
@@ -447,6 +457,11 @@ defmodule Decimal do
   def div(%Decimal{coef: 0}, %Decimal{coef: 0}),
     do: error(:invalid_operation, "0 / 0", %Decimal{coef: :NaN})
 
+  def div(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
+      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
+    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  end
+
   def div(%Decimal{sign: sign1}, %Decimal{sign: sign2, coef: 0}) do
     sign = if sign1 == sign2, do: 1, else: -1
     error(:division_by_zero, nil, %Decimal{sign: sign, coef: :inf})
@@ -518,6 +533,11 @@ defmodule Decimal do
     error(:division_by_zero, nil, %Decimal{sign: div_sign, coef: :inf})
   end
 
+  def div_int(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
+      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
+    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  end
+
   def div_int(%Decimal{} = num1, %Decimal{} = num2) do
     %Decimal{sign: sign1, coef: coef1, exp: exp1} = num1
     %Decimal{sign: sign2, coef: coef2, exp: exp2} = num2
@@ -582,6 +602,11 @@ defmodule Decimal do
 
   def rem(%Decimal{sign: sign1}, %Decimal{coef: 0}),
     do: error(:division_by_zero, nil, %Decimal{sign: sign1, coef: 0})
+
+  def rem(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
+      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
+    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  end
 
   def rem(%Decimal{} = num1, %Decimal{} = num2) do
     %Decimal{sign: sign1, coef: coef1, exp: exp1} = num1
@@ -663,6 +688,12 @@ defmodule Decimal do
     div_error = error(:division_by_zero, nil, %Decimal{sign: div_sign, coef: :inf})
     rem_error = error(:division_by_zero, nil, %Decimal{sign: sign1, coef: 0})
     {div_error, rem_error}
+  end
+
+  def div_rem(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
+      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
+    error = error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+    {error, error}
   end
 
   def div_rem(%Decimal{} = num1, %Decimal{} = num2) do
@@ -863,6 +894,11 @@ defmodule Decimal do
   def mult(%Decimal{coef: :inf}, %Decimal{coef: 0}),
     do: error(:invalid_operation, "0 * ±Infinity", %Decimal{coef: :NaN})
 
+  def mult(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
+      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
+    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  end
+
   def mult(%Decimal{sign: sign1, coef: :inf, exp: exp1}, %Decimal{sign: sign2, exp: exp2}) do
     sign = if sign1 == sign2, do: 1, else: -1
     # exponent?
@@ -908,12 +944,19 @@ defmodule Decimal do
     %{num | exp: 0}
   end
 
+  def normalize(%Decimal{sign: sign, coef: coef})
+      when coef < 0 or sign not in [1, -1] do
+    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
+      coef: :NaN
+    })
+  end
+
+  def normalize(%Decimal{sign: sign, coef: 0, exp: _exp}) do
+    %Decimal{sign: sign, coef: 0, exp: 0}
+  end
+
   def normalize(%Decimal{sign: sign, coef: coef, exp: exp}) do
-    if coef == 0 do
-      %Decimal{sign: sign, coef: 0, exp: 0}
-    else
-      %{do_normalize(coef, exp) | sign: sign} |> context
-    end
+    %{do_normalize(coef, exp) | sign: sign} |> context
   end
 
   @doc """
@@ -938,6 +981,13 @@ defmodule Decimal do
   def round(%Decimal{coef: :NaN} = num, _, _), do: num
 
   def round(%Decimal{coef: :inf} = num, _, _), do: num
+
+  def round(%Decimal{sign: sign, coef: coef}, _, _)
+      when coef < 0 or sign not in [1, -1] do
+    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
+      coef: :NaN
+    })
+  end
 
   def round(%Decimal{} = num, n, mode) do
     %Decimal{sign: sign, coef: coef, exp: exp} = normalize(num)
@@ -967,6 +1017,13 @@ defmodule Decimal do
 
   def sqrt(%Decimal{coef: 0, exp: exp} = num),
     do: %{num | exp: exp >>> 1}
+
+  def sqrt(%Decimal{sign: sign, coef: coef})
+      when coef < 0 or sign not in [1, -1] do
+    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
+      coef: :NaN
+    })
+  end
 
   def sqrt(%Decimal{sign: -1} = num),
     do: error(:invalid_operation, "less than zero", num)
@@ -1080,7 +1137,11 @@ defmodule Decimal do
       #Decimal<3.14>
   """
   @spec new(decimal) :: t
-  def new(%Decimal{} = num), do: num
+  def new(%Decimal{sign: sign, coef: coef, exp: exp} = num)
+      when sign in [1, -1] and ((is_integer(coef) and coef >= 0) or coef in [:NaN, :inf]) and
+             is_integer(exp) do
+    num
+  end
 
   def new(int) when is_integer(int),
     do: %Decimal{sign: if(int < 0, do: -1, else: 1), coef: Kernel.abs(int)}
@@ -1090,6 +1151,10 @@ defmodule Decimal do
       {decimal, ""} -> decimal
       _ -> raise Error, reason: "number parsing syntax: #{inspect(binary)}"
     end
+  end
+
+  def new(num) do
+    raise Error, reason: "wrong decimal number: (#{inspect(num)})"
   end
 
   @doc """
@@ -1159,6 +1224,12 @@ defmodule Decimal do
   """
   @spec cast(term) :: {:ok, t} | :error
   def cast(integer) when is_integer(integer), do: {:ok, Decimal.new(integer)}
+
+  def cast(%Decimal{sign: sign, coef: coef, exp: exp})
+      when coef < 0 or sign not in [1, -1] or not is_integer(exp) do
+    :error
+  end
+
   def cast(%Decimal{} = decimal), do: {:ok, decimal}
   def cast(float) when is_float(float), do: {:ok, from_float(float)}
 
@@ -1228,6 +1299,13 @@ defmodule Decimal do
 
   def to_string(%Decimal{sign: sign, coef: :inf}, _type) do
     if sign == 1, do: "Infinity", else: "-Infinity"
+  end
+
+  def to_string(%Decimal{sign: sign, coef: coef}, _type)
+      when coef < 0 or sign not in [1, -1] do
+    error(:invalid_operation, "coefficient (#{coef}) or sign (#{sign}) are wrong", %Decimal{
+      coef: :NaN
+    })
   end
 
   def to_string(%Decimal{sign: sign, coef: coef, exp: exp}, :normal) do
@@ -1316,6 +1394,14 @@ defmodule Decimal do
   Fails when loss of precision will occur.
   """
   @spec to_integer(t) :: integer
+
+  def to_integer(%Decimal{sign: sign, coef: coef})
+      when coef < 0 or sign not in [1, -1] do
+    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
+      coef: :NaN
+    })
+  end
+
   def to_integer(%Decimal{sign: sign, coef: coef, exp: 0})
       when is_integer(coef),
       do: sign * coef
@@ -1335,6 +1421,13 @@ defmodule Decimal do
   the decimal cannot be converted to a float.
   """
   @spec to_float(t) :: float
+  def to_float(%Decimal{sign: sign, coef: coef})
+      when coef < 0 or sign not in [1, -1] do
+    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
+      coef: :NaN
+    })
+  end
+
   def to_float(%Decimal{sign: sign, coef: coef, exp: exp}) when is_integer(coef) do
     # Convert back to float without loss
     # http://www.exploringbinary.com/correct-decimal-to-floating-point-using-big-integers/