--- a/lib/decimal.ex
+++ b/lib/decimal.ex
@@ -229,9 +229,19 @@ defmodule Decimal do
 
   def add(%Decimal{}, %Decimal{coef: :inf} = num2), do: num2
 
-  def add(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
-      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
-    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  def add(%Decimal{coef: coef1}, %Decimal{coef: coef2})
+      when coef1 < 0 or coef2 < 0 do
+    error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
+  end
+
+  def add(%Decimal{sign: sign1}, %Decimal{})
+      when sign1 != 1 and sign1 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
+  end
+
+  def add(%Decimal{}, %Decimal{sign: sign2})
+      when sign2 != 1 and sign2 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def add(%Decimal{} = num1, %Decimal{} = num2) do
@@ -314,9 +324,19 @@ defmodule Decimal do
   def compare(_num1, %Decimal{coef: :NaN} = num2),
     do: error(:invalid_operation, "operation on NaN", num2)
 
-  def compare(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
-      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
-    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  def compare(%Decimal{coef: coef1}, %Decimal{coef: coef2})
+      when coef1 < 0 or coef2 < 0 do
+    error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
+  end
+
+  def compare(%Decimal{sign: sign1}, %Decimal{})
+      when sign1 != 1 and sign1 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
+  end
+
+  def compare(%Decimal{}, %Decimal{sign: sign2})
+      when sign2 != 1 and sign2 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def compare(%Decimal{} = num1, %Decimal{} = num2) do
@@ -457,9 +477,19 @@ defmodule Decimal do
   def div(%Decimal{coef: 0}, %Decimal{coef: 0}),
     do: error(:invalid_operation, "0 / 0", %Decimal{coef: :NaN})
 
-  def div(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
-      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
-    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  def div(%Decimal{coef: coef1}, %Decimal{coef: coef2})
+      when coef1 < 0 or coef2 < 0 do
+    error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
+  end
+
+  def div(%Decimal{sign: sign1}, %Decimal{})
+      when sign1 != 1 and sign1 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
+  end
+
+  def div(%Decimal{}, %Decimal{sign: sign2})
+      when sign2 != 1 and sign2 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def div(%Decimal{sign: sign1}, %Decimal{sign: sign2, coef: 0}) do
@@ -533,9 +563,19 @@ defmodule Decimal do
     error(:division_by_zero, nil, %Decimal{sign: div_sign, coef: :inf})
   end
 
-  def div_int(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
-      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
-    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  def div_int(%Decimal{coef: coef1}, %Decimal{coef: coef2})
+      when coef1 < 0 or coef2 < 0 do
+    error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
+  end
+
+  def div_int(%Decimal{sign: sign1}, %Decimal{})
+      when sign1 != 1 and sign1 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
+  end
+
+  def div_int(%Decimal{}, %Decimal{sign: sign2})
+      when sign2 != 1 and sign2 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def div_int(%Decimal{} = num1, %Decimal{} = num2) do
@@ -603,9 +643,19 @@ defmodule Decimal do
   def rem(%Decimal{sign: sign1}, %Decimal{coef: 0}),
     do: error(:division_by_zero, nil, %Decimal{sign: sign1, coef: 0})
 
-  def rem(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
-      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
-    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  def rem(%Decimal{coef: coef1}, %Decimal{coef: coef2})
+      when coef1 < 0 or coef2 < 0 do
+    error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
+  end
+
+  def rem(%Decimal{sign: sign1}, %Decimal{})
+      when sign1 != 1 and sign1 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
+  end
+
+  def rem(%Decimal{}, %Decimal{sign: sign2})
+      when sign2 != 1 and sign2 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def rem(%Decimal{} = num1, %Decimal{} = num2) do
@@ -690,9 +740,21 @@ defmodule Decimal do
     {div_error, rem_error}
   end
 
-  def div_rem(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
-      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
-    error = error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  def div_rem(%Decimal{coef: coef1}, %Decimal{coef: coef2})
+      when coef1 < 0 or coef2 < 0 do
+    error = error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
+    {error, error}
+  end
+
+  def div_rem(%Decimal{sign: sign1}, %Decimal{})
+      when sign1 != 1 and sign1 != -1 do
+    error = error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
+    {error, error}
+  end
+
+  def div_rem(%Decimal{}, %Decimal{sign: sign2})
+      when sign2 != 1 and sign2 != -1 do
+    error = error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
     {error, error}
   end
 
@@ -894,9 +956,19 @@ defmodule Decimal do
   def mult(%Decimal{coef: :inf}, %Decimal{coef: 0}),
     do: error(:invalid_operation, "0 * ±Infinity", %Decimal{coef: :NaN})
 
-  def mult(%Decimal{sign: sign1, coef: coef1}, %Decimal{sign: sign2, coef: coef2})
-      when coef1 < 0 or coef2 < 0 or sign1 not in [1, -1] or sign2 not in [1, -1] do
-    error(:invalid_operation, "wrong sign or coef", %Decimal{coef: :NaN})
+  def mult(%Decimal{coef: coef1}, %Decimal{coef: coef2})
+      when coef1 < 0 or coef2 < 0 do
+    error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
+  end
+
+  def mult(%Decimal{sign: sign1}, %Decimal{})
+      when sign1 != 1 and sign1 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
+  end
+
+  def mult(%Decimal{}, %Decimal{sign: sign2})
+      when sign2 != 1 and sign2 != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def mult(%Decimal{sign: sign1, coef: :inf, exp: exp1}, %Decimal{sign: sign2, exp: exp2}) do
@@ -944,11 +1016,14 @@ defmodule Decimal do
     %{num | exp: 0}
   end
 
-  def normalize(%Decimal{sign: sign, coef: coef})
-      when coef < 0 or sign not in [1, -1] do
-    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
-      coef: :NaN
-    })
+  def normalize(%Decimal{coef: coef})
+      when coef < 0 do
+    error(:invalid_operation, "wrong coefficient", %Decimal{coef: :NaN})
+  end
+
+  def normalize(%Decimal{sign: sign})
+      when sign != 1 and sign != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def normalize(%Decimal{sign: sign, coef: 0, exp: _exp}) do
@@ -982,11 +1057,14 @@ defmodule Decimal do
 
   def round(%Decimal{coef: :inf} = num, _, _), do: num
 
-  def round(%Decimal{sign: sign, coef: coef}, _, _)
-      when coef < 0 or sign not in [1, -1] do
-    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
-      coef: :NaN
-    })
+  def round(%Decimal{coef: coef}, _, _)
+      when coef < 0 do
+    error(:invalid_operation, "wrong coefficient", %Decimal{coef: :NaN})
+  end
+
+  def round(%Decimal{sign: sign}, _, _)
+      when sign != 1 and sign != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def round(%Decimal{} = num, n, mode) do
@@ -1018,11 +1096,14 @@ defmodule Decimal do
   def sqrt(%Decimal{coef: 0, exp: exp} = num),
     do: %{num | exp: exp >>> 1}
 
-  def sqrt(%Decimal{sign: sign, coef: coef})
-      when coef < 0 or sign not in [1, -1] do
-    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
-      coef: :NaN
-    })
+  def sqrt(%Decimal{coef: coef})
+      when coef < 0 do
+    error(:invalid_operation, "wrong coefficient", %Decimal{coef: :NaN})
+  end
+
+  def sqrt(%Decimal{sign: sign})
+      when sign != 1 and sign != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def sqrt(%Decimal{sign: -1} = num),
@@ -1225,8 +1306,18 @@ defmodule Decimal do
   @spec cast(term) :: {:ok, t} | :error
   def cast(integer) when is_integer(integer), do: {:ok, Decimal.new(integer)}
 
-  def cast(%Decimal{sign: sign, coef: coef, exp: exp})
-      when coef < 0 or sign not in [1, -1] or not is_integer(exp) do
+  def cast(%Decimal{coef: coef})
+      when coef < 0 do
+    :error
+  end
+
+  def cast(%Decimal{sign: sign})
+      when sign != 1 and sign != -1 do
+    :error
+  end
+
+  def cast(%Decimal{exp: exp})
+      when not is_integer(exp) do
     :error
   end
 
@@ -1301,11 +1392,14 @@ defmodule Decimal do
     if sign == 1, do: "Infinity", else: "-Infinity"
   end
 
-  def to_string(%Decimal{sign: sign, coef: coef}, _type)
-      when coef < 0 or sign not in [1, -1] do
-    error(:invalid_operation, "coefficient (#{coef}) or sign (#{sign}) are wrong", %Decimal{
-      coef: :NaN
-    })
+  def to_string(%Decimal{coef: coef}, _type)
+      when coef < 0 do
+    error(:invalid_operation, "wrong coefficient", %Decimal{coef: :NaN})
+  end
+
+  def to_string(%Decimal{sign: sign}, _type)
+      when sign != 1 and sign != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def to_string(%Decimal{sign: sign, coef: coef, exp: exp}, :normal) do
@@ -1394,12 +1488,14 @@ defmodule Decimal do
   Fails when loss of precision will occur.
   """
   @spec to_integer(t) :: integer
+  def to_integer(%Decimal{coef: coef})
+      when coef < 0 do
+    error(:invalid_operation, "wrong coefficient", %Decimal{coef: :NaN})
+  end
 
-  def to_integer(%Decimal{sign: sign, coef: coef})
-      when coef < 0 or sign not in [1, -1] do
-    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
-      coef: :NaN
-    })
+  def to_integer(%Decimal{sign: sign})
+      when sign != 1 and sign != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def to_integer(%Decimal{sign: sign, coef: coef, exp: 0})
@@ -1421,11 +1517,14 @@ defmodule Decimal do
   the decimal cannot be converted to a float.
   """
   @spec to_float(t) :: float
-  def to_float(%Decimal{sign: sign, coef: coef})
-      when coef < 0 or sign not in [1, -1] do
-    error(:invalid_operation, "wrong coefficient (#{coef}) or sign (#{sign}) ", %Decimal{
-      coef: :NaN
-    })
+  def to_float(%Decimal{coef: coef})
+      when coef < 0 do
+    error(:invalid_operation, "wrong coefficient", %Decimal{coef: :NaN})
+  end
+
+  def to_float(%Decimal{sign: sign})
+      when sign != 1 and sign != -1 do
+    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
   end
 
   def to_float(%Decimal{sign: sign, coef: coef, exp: exp}) when is_integer(coef) do