Hi,

I found a bug or structural problem in Decimal what could have some big impact under special circumstances. Here an example from the Money lib where Decimal is used as main backbone:

```elixir
iex(15)> m = Money.new(:USD, Decimal.new(%Decimal{sign: -1, coef: -12000000000}))
#Money<:USD, --12000000000>
iex(16)> Money.mult(m, 3)
{:ok, #Money<:USD, --36000000000>} 
iex(17)> Money.mult!(m, 3 )|> Money.to_string 
{:ok, "$--*,000,000,000.00"}
```

Of course it's a special edge case ... normally the best practice is:

```elixir
iex(18)> m = Money.new(:USD, "-12000000000") 
#Money<:USD, -12000000000>
```
but you never know ... maybe someone using it ...it's possible! it's valid Code ! 
NO compiler warning! (also true if you are using dialyzer, when using var names for input)
NO runtime error!

And the worst case, if you are using any division with this kind of Decimal number like this:

```elixir
iex(19)> m = Money.new(:USD, Decimal.new(%Decimal{sign: -1, coef: -12000000000}))
#Money<:USD, --12000000000>
iex(20)> Money.mult!(m, 3) |> Money.to_decimal |> Decimal.to_float
```
-> you end up in an infinite loop!!! 

The reason for this is clear: the %Decimal{} struct - with not enough validation for input -> so you can input a negative coefficient.

The type spec and docs are perfect - but do not prevent any misuse: 
Type spec :  
 @type coefficient :: non_neg_integer | :NaN | :inf
  @type sign :: 1 | -1

Documentation: The coefficient of the power of `10`. Non-negative because the sign is stored separately in `sign`.

So my suggestion for a solution you can find in this pull request: mostly checks for this special edge case.

And this time I included the corresponding tests. ;-) 

one example for the additional function clause (now backwards compatible to Elixir 1.2):
```elixir
def add(%Decimal{coef: coef1}, %Decimal{coef: coef2})
      when coef1 < 0 or coef2 < 0 do
    error(:invalid_operation, "wrong coef", %Decimal{coef: :NaN})
  end

  def add(%Decimal{sign: sign1}, %Decimal{})
      when sign1 != 1 and sign1 != -1 do
    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
  end

  def add(%Decimal{}, %Decimal{sign: sign2})
      when sign2 != 1 and sign2 != -1 do
    error(:invalid_operation, "wrong sign", %Decimal{coef: :NaN})
  end
```

The test and benchees are fine - no big performance cost - the differences are marginal (see deviation):

before the fix:

```elixir
##### With input 30 Digits Integer #####
Name                           ips        average  deviation         median         99th %
Decimal.add()    Int      318.01 K        3.14 μs   ±771.75%           3 μs           5 μs

Memory usage statistics:

Name                    Memory usage
Decimal.add()    Int         4.26 KB

**All measurements for memory usage were the same**

```

with the fix:
```elixir
##### With input 30 Digits Integer #####
Name                           ips        average  deviation         median         99th %
Decimal.add()    Int      316.97 K        3.15 μs   ±659.73%           3 μs           5 μs

Memory usage statistics:

Name                    Memory usage
Decimal.add()    Int         4.26 KB

**All measurements for memory usage were the same**

```