@ericmj 2020-06-22
It would be hard to catch every scenario where a user constructs an invalid `Decimal` struct. They could for example pass a float in any of the fields and the library would also work incorrectly. Adding guards to every function that catches every invalid value seems fragile and not very feasible. In addition we are catching the error too late, at the time where the struct is passed to the function instead of when the invalid struct is created.

It is better to use the provided functions such as `Decimal.new/1` to create decimals than to create them manually. In the past the `Decimal.t/0` type was opaque to prevent users from creating structs manually, but libraries such as Ecto/Postgrex needed to create structs manually when casting the values from postgres' wire format. I think that's one of the only times where you should not (cannot) create the structs manually.

I appreciate the PR, but because of these reasons I am hesitant to merge it.