@wojtekmach 2020-06-22
We already have an `is_decimal/1` macro, I believe it would be possible to extend to check the struct more closely (e.g. `%Decimal{coef: coef} when is_integer(coef) and coef >= 0` and the equivalent for guards) and use it for existing functions.

That being said, as mentioned in this issue and [elixirforum](https://elixirforum.com/t/decimal-bug-please-be-careful-with-negative-coefficients-fix-is-on-pull-request/32430), it is more common to assume that a given struct was properly constructed in the first place and skip further checks (that's why they are _structs_ and not merely _maps_). I believe extending the is_decimal/1 macro further would have small additional runtime cost too but perhaps it's negligible anyway.