@razuf 2020-06-22
Many thanks for reviewing PR and for your comment. I appreciate that.

I can follow that argument:

> Adding guards to every function that catches every invalid value seems fragile and not very feasible.

and too: 
> we are catching the error too late

thats why I had the check in new/1 and cast/1 - but what do you think about the adapted PR with only checking that (I removed the rest): 

For the inbound:
new/1 - is the same check as in new/3 
```elixir
  def new(%Decimal{sign: sign, coef: coef, exp: exp} = num)
      when sign in [1, -1] and ((is_integer(coef) and coef >= 0) or coef in [:NaN, :inf]) and
             is_integer(exp) do
    num
  end
```

and added a default case error (in my opinion better error handling then FunctionClauseError - but I can live without that..)

```elixir
  def new(num) do
    raise Error, reason: "wrong decimal number: #{inspect(num)}"
  end

```

cast - I think that's the behaviour from cast I would expect - almost always :error instead of "let it crash":

For the outbound:
to_string, to_integer, to_float

That's fixing most of the tests in the decimal_test.ex from before - and increase the security of the Decimal lib - I think" ;-)

one of the edge case did not fixed would be only:

```elixir
test "test sqrt with wrong sign -> infinite loop" do
  assert_raise Decimal.Error, fn ->
    Decimal.sqrt(d(3, 1, -1))
  end
end

```

As you said - we cannot handle all possible cases...

But the Money examples from above would be fixed by this PR.